requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
umap-learn
openpyxl
h5py
scipy
librosa
soundfile
